perf: tileset comparison to speed up move validity checks
`TileSet` is now represented as a bitmask (u16) where
each bit represents the presence of a tile. And `Board`
now has a `played_tiles: TileSet` field that tracks
the tiles that have already been played. With both
of these changes, we can speed up drastically the
`is_move_possible` method  (from  1.6μs to 0.5ns).

This comes with the drawback of having to update
the `played_tiles` field each time we play a move.